+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/GameOverMenuState.java
@@ -1,6 +1,9 @@
 package com.github.mjkuranda.spaceadventure2.states.menus;
 
 import com.github.mjkuranda.spaceadventure2.PlayerData;
+import com.github.mjkuranda.spaceadventure2.entities.EntityShapes;
+import com.github.mjkuranda.spaceadventure2.entities.Particle;
+import com.github.mjkuranda.spaceadventure2.resources.GameAnimation;
 import com.github.mjkuranda.spaceadventure2.resources.GameFont;
 import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 import com.github.mjkuranda.spaceadventure2.states.StatesId;
@@ -12,16 +15,24 @@ import org.newdawn.slick.GameContainer;
 import org.newdawn.slick.Graphics;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
+import org.newdawn.slick.geom.Shape;
 import org.newdawn.slick.state.StateBasedGame;
+
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Random;
 
 public class GameOverMenuState extends MenuState {
 
     private float approximation, speed = 4.0f;
     private boolean isApproximating = true;
     private byte currentBackgroundImageIndex = 0;
+    private List<Particle> particleList;
 
     public GameOverMenuState(StateBasedGame game) {
         super();
+
+        this.particleList = new LinkedList<>();
 
         this.bindOptions(new MenuOption[] {
                 new SimpleMenuOption("Play again")
@@ -38,6 +49,13 @@ public class GameOverMenuState extends MenuState {
     @Override
     public int getID() {
         return StatesId.GAME_OVER_MENU;
+    }
+
+    @Override
+    public void render(GameContainer container, StateBasedGame game, Graphics g) throws SlickException {
+        super.render(container, game, g);
+
+        renderParticles();
     }
 
     @Override
@@ -64,9 +82,22 @@ public class GameOverMenuState extends MenuState {
         g.drawString(str, x, y);
     }
 
+    private void renderParticles() {
+        for (var particle : particleList) {
+            float x = particle.getX();
+            float y = particle.getY();
+
+            particle.getAnimation().draw(x, y, 64, 64);
+        }
+    }
+
     @Override
     public void update(GameContainer container, StateBasedGame game, int delta) throws SlickException {
         super.update(container, game, delta);
+
+        if (particleList.size() > 0 && particleList.get(0).getAnimation().isStopped()) {
+            particleList.clear();
+        }
 
         if(approximation >= -128 && approximation <= 1080) speed = 3.0f;
         if(approximation >= -200 && approximation <= -127) speed = 3.5f;
@@ -98,5 +129,27 @@ public class GameOverMenuState extends MenuState {
             case 3 -> GameImage.BACKGROUND_2;
             default -> throw new IllegalStateException("Unexpected value: " + currentBackgroundImageIndex);
         };
+    }
+
+    @Override
+    public void init(GameContainer container, StateBasedGame game) {
+        if (GameAnimation.EXPLOSION == null) {
+            return;
+        }
+
+        Random r = new Random();
+
+        for (int i = 0; i < 6400; i++) {
+            float x = r.nextFloat() * (container.getWidth() + 256) - 128;
+            float y = r.nextFloat() * (container.getHeight() + 256) - 128;
+            Shape shape = switch (r.nextInt(3)) {
+                case 0 -> EntityShapes.SMALL_EXPLOSION_SHAPE;
+                case 1 -> EntityShapes.MEDIUM_EXPLOSION_SHAPE;
+                case 2 -> EntityShapes.LARGE_EXPLOSION_SHAPE;
+                default -> throw new IllegalStateException("Unexpected value: " + r.nextInt(3));
+            };
+
+            particleList.add(new Particle(GameAnimation.EXPLOSION, x, y, shape));
+        }
     }
 }