+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/MainMenuState.java
@@ -1,24 +1,34 @@
 package com.github.mjkuranda.spaceadventure2.states.menus;
 
 import com.github.mjkuranda.spaceadventure2.SpaceAdventure2;
+import com.github.mjkuranda.spaceadventure2.entities.EntityShapes;
+import com.github.mjkuranda.spaceadventure2.entities.Particle;
+import com.github.mjkuranda.spaceadventure2.resources.GameAnimation;
 import com.github.mjkuranda.spaceadventure2.resources.GameFont;
 import com.github.mjkuranda.spaceadventure2.states.StatesId;
 import com.github.mjkuranda.spaceadventure2.states.menus.options.MenuOption;
 import com.github.mjkuranda.spaceadventure2.states.menus.options.SimpleMenuOption;
 import com.github.mjkuranda.spaceadventure2.states.menus.options.events.EnterStateMenuOptionEvent;
 import com.github.mjkuranda.spaceadventure2.states.menus.options.events.ExitGameMenuOptionEvent;
-import org.newdawn.slick.Color;
-import org.newdawn.slick.GameContainer;
-import org.newdawn.slick.Graphics;
-import org.newdawn.slick.UnicodeFont;
+import org.newdawn.slick.*;
+import org.newdawn.slick.geom.Shape;
 import org.newdawn.slick.state.StateBasedGame;
 
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Random;
+
 public class MainMenuState extends MenuState {
+
+    /** Particle list for explosions */
+    private List<Particle> particleList;
 
     private BlindingLabel navigationInfoLabel;
 
     public MainMenuState(StateBasedGame game) {
         super();
+
+        this.particleList = new LinkedList<>();
 
         this.navigationInfoLabel = new BlindingLabel("Press up or down arrow to navigate and ENTER to select");
 
@@ -34,6 +44,43 @@ public class MainMenuState extends MenuState {
                 new SimpleMenuOption("Exit")
                         .bindOnSelectEvent(new ExitGameMenuOptionEvent())
         });
+    }
+
+    @Override
+    public void init(GameContainer container, StateBasedGame game) {
+        if (GameAnimation.EXPLOSION == null) {
+            return;
+        }
+
+        Random r = new Random();
+
+        for (int i = 0; i < 6400; i++) {
+            float x = r.nextFloat() * (container.getWidth() + 256) - 128;
+            float y = r.nextFloat() * (container.getHeight() + 256) - 128;
+            Shape shape = switch (r.nextInt(3)) {
+                case 0 -> EntityShapes.SMALL_EXPLOSION_SHAPE;
+                case 1 -> EntityShapes.MEDIUM_EXPLOSION_SHAPE;
+                case 2 -> EntityShapes.LARGE_EXPLOSION_SHAPE;
+                default -> throw new IllegalStateException("Unexpected value: " + r.nextInt(3));
+            };
+
+            particleList.add(new Particle(GameAnimation.EXPLOSION, x, y, shape));
+        }
+    }
+
+    @Override
+    public void render(GameContainer container, StateBasedGame game, Graphics g) throws SlickException {
+        super.render(container, game, g);
+
+        renderParticles();
+    }
+
+    public void update(GameContainer container, StateBasedGame game, int delta) throws SlickException {
+        super.update(container, game, delta);
+
+        if (particleList.size() > 0 && particleList.get(0).getAnimation().isStopped()) {
+            particleList.clear();
+        }
     }
 
     @Override
@@ -69,6 +116,15 @@ public class MainMenuState extends MenuState {
         // Navigation text
         float navX = container.getWidth() / 2 - navigationInfoLabel.getTextWidth() / 2;
         navigationInfoLabel.render(g, navX, container.getHeight() - 128);
+    }
+
+    private void renderParticles() {
+        for (var particle : particleList) {
+            float x = particle.getX();
+            float y = particle.getY();
+
+            particle.getAnimation().draw(x, y, 64, 64);
+        }
     }
 
     private int getStringWidth(String str, UnicodeFont font) {