+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/menus/options/events/EnterStateMenuOptionEvent.java
@@ -18,6 +18,7 @@ public class EnterStateMenuOptionEvent implements MenuOptionEvent {
 
     @Override
     public void onSelect() {
+        game.getContainer().getInput().clearKeyPressedRecord();
         game.enterState(stateId);
         menu.reset();
     }